test-otel-kafka/pyproject.toml
[project]
name = "test"
version = "0"
requires-python = ">=3.13"

dependencies = [
    "confluent-kafka>=2.9.0",
    "ipdb>=0.13.13",
    "sentry-sdk",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-instrumentation-confluent-kafka",
    "opentelemetry-exporter-otlp",
]

[tool.uv.sources]
sentry-sdk = { path = "../../sentry-python", editable = true }
